feat: Update CD pipeline with Docker build and push
This commit updates the Continuous Deployment (CD) pipeline by replacing
the Go binary build step with a Docker image build step. Additionally,
it introduces a new step to push the built Docker image to the Google
Cloud Platform (GCP) Artifact Registry. This modification is crucial
for deploying the application in a containerized environment.