feat: add GCP auth and setup in CD workflow
This commit introduces the use of google-github-actions for
authentication and setup of the GCP SDK in the Continuous Deployment
workflow. This change ensures that the GCP credentials are securely
used and the GCP SDK is properly set up before building and pushing
the Docker image to the GCP Artifact Registry.